+++ coderd/coderd.go
@@ -25,10 +25,10 @@ import (
 	"github.com/coder/coder/coderd/httpapi"
 	"github.com/coder/coder/coderd/httpmw"
 	"github.com/coder/coder/coderd/rbac"
+	"github.com/coder/coder/coderd/tracing"
 	"github.com/coder/coder/coderd/turnconn"
 	"github.com/coder/coder/codersdk"
 	"github.com/coder/coder/site"
-	"github.com/coder/coder/tracing"
 )
 
 // Options are requires parameters for Coder to start.
